# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: MyWeb E-commerce API
  version: "1.0.0"
  description: |
    REST API for MyWeb ecommerce (clothes, dresses, shoes).
    
    **Auth**: JWT Bearer. Public endpoints: `/api/v1/health`, `/api/v1/auth/**`, `/api/v1/categories/**`, `/api/v1/products/**`.
    All others require JWT. Admin endpoints additionally require `ROLE_ADMIN`.
servers:
  - url: http://localhost:8080
    description: Local dev
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      in: query
      name: page
      schema:
        type: integer
        minimum: 0
        default: 0
      description: 0-based page index
    SizeParam:
      in: query
      name: size
      schema:
        type: integer
        minimum: 1
        default: 20
      description: Page size
    ProductSortParam:
      in: query
      name: sortBy
      schema:
        type: string
        enum: [newest, oldest, priceAsc, priceDesc]
        default: newest
      description: Sort option
    ProductKeywordParam:
      in: query
      name: keyword
      schema:
        type: string
      description: Full-text keyword for products
    ProductCategoryParam:
      in: query
      name: category
      schema:
        type: string
      description: Category slug filter
...
  responses:
    Unauthorized:
      description: Missing/invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Not allowed (e.g., missing ADMIN role)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string

    # ---- Auth / User ----
...
    AdminProductResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        slug:
          type: string
        shortDesc:
          type: string
        longDesc:
          type: string
        categoryId:
          type: integer
          format: int64
        categoryName:
          type: string
        categorySlug:
          type: string
        basePrice:
          type: number
          format: double
        stockQuantity:
          type: integer
        thumbnailUrl:
          type: string
        status:
          type: string
          enum: [ACTIVE, INACTIVE, DRAFT]
        imageUrls:
          type: array
          items:
            type: string
          description: List of additional product image URLs
...
    ProductListItemResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        slug:
          type: string
        thumbnailUrl:
          type: string
        price:
          type: number
          format: double
        categorySlug:
          type: string
        categoryName:
          type: string
        createdAt:
          type: string
          format: date-time
...
    ReviewRequest:
      type: object
      required: [orderItemId, rating]
      properties:
        orderItemId:
          type: integer
          format: int64
          description: Order item ID to link review with specific purchase (required for verification)
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        imageUrl:
          type: string
    ReviewResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        username:
          type: string
        userFullName:
          type: string
          description: Kept for backward compatibility
        productId:
          type: integer
          format: int64
        rating:
          type: integer
        comment:
          type: string
        imageUrl:
          type: string
        # Edit tracking (Shopee-level feature)
        editCount:
          type: integer
          description: Number of times review has been edited (max 2)
        isEdited:
          type: boolean
          description: Whether review has been edited
        # Purchase verification
        orderId:
          type: integer
          format: int64
          nullable: true
        orderNumber:
          type: string
          nullable: true
        verifiedPurchase:
          type: boolean
          description: True if user actually purchased this product (Verified Purchase badge)
        adminReply:
          type: string
        adminRepliedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AdminReplyRequest:
      type: object
      required: [reply]
      properties:
        reply:
          type: string

    # ---- Dashboard ----
...
    Page_AdminProductResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/AdminProductResponse'
        number:
          type: integer
          example: 0
        size:
          type: integer
          example: 20
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
    Page_ProductListItemResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ProductListItemResponse'
        number:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
...
paths:
...
  /api/v1/products:
    get:
      tags:
        - "CLIENT - Products & Catalog"
      security: []
      summary: List products (public)
      parameters:
        - $ref: '#/components/parameters/ProductCategoryParam'
        - $ref: '#/components/parameters/ProductKeywordParam'
        - in: query
          name: minPrice
          schema:
            type: number
            format: double
          description: Minimum price filter
        - in: query
          name: maxPrice
          schema:
            type: number
            format: double
          description: Maximum price filter
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/ProductSortParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_ProductListItemResponse'
...
  /api/v1/products/{productId}/reviews:
    get:
      tags:
        - "CLIENT - Reviews"
      security: []
      summary: List reviews for product
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ReviewResponse' }
    post:
      tags:
        - "CLIENT - Reviews"
      summary: Add review (User)
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReviewRequest' }
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReviewResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/products/reviews/{reviewId}/me:
    put:
      tags:
        - "CLIENT - Reviews"
      summary: Update own review (max 2 edits allowed - Shopee rule)
      description: |
        Edit your own review. Maximum 2 edits allowed per review.
        Must be the review owner.
      parameters:
        - in: path
          name: reviewId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewRequest'
      responses:
        '200':
          description: Review updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
        '400':
          description: Maximum edit limit reached (2 times)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Not your review
        '404':
          description: Review not found
    delete:
      tags:
        - "CLIENT - Reviews"
      summary: Delete my review (User)
      parameters:
        - in: path
          name: reviewId
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '204': { description: No Content }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/products/reviews/{reviewId}/reply:
    post:
      tags:
        - "ADMIN - Review Moderation"
      summary: Reply to review (ADMIN)
      parameters:
        - in: path
          name: reviewId
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminReplyRequest' }
      responses:
        '200':
          description: Replied
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReviewResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/products/reviews/{reviewId}:
    delete:
      tags:
        - "ADMIN - Review Moderation"
      summary: Delete review (ADMIN)
      parameters:
        - in: path
          name: reviewId
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '204': { description: No Content }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ---- File Upload ----
...
  /api/v1/admin/reviews:
    get:
      tags:
        - "ADMIN - Review Moderation"
      summary: Get all reviews (ADMIN)
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 0 }
        - in: query
          name: size
          schema: { type: integer, default: 20 }
        - in: query
          name: rating
          schema: { type: integer }
          description: Optional rating filter (1-5)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_AdminProductResponse' # Reuse page wrapper structure or similar
  /api/v1/admin/reviews/{id}/visibility:
    put:
      tags:
        - "ADMIN - Review Moderation"
      summary: Toggle review visibility (ADMIN)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
        - in: query
          name: visible
          required: true
          schema: { type: boolean }
      responses:
        '204': { description: No Content }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---- Admin Vouchers ----
...